# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MarkHub API
  version: 0.1.0
  description: 'Self-hosted bookmark hub — unified contract for Docker (FastAPI) and
    Cloudflare Workers. Error envelope: `{ "error": { "code", "message", "details?"
    } }`.'
servers:
- url: /
paths:
...
  /api/v1/backup/export:
    get:
      tags:
      - backup
      summary: Export
      operationId: export_api_v1_backup_export_get
      security:
      - HTTPBearer: []
      parameters:
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/backup/import:
    post:
      tags:
      - backup
      summary: Import Json
      operationId: import_json_api_v1_backup_import_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__api__backup__ImportBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/backup/import-file:
    post:
      tags:
      - backup
      summary: Import File
      operationId: import_file_api_v1_backup_import_file_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_file_api_v1_backup_import_file_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/backup/webdav:
    get:
      tags:
      - backup
      summary: Get Webdav
      operationId: get_webdav_api_v1_backup_webdav_get
      security:
      - HTTPBearer: []
      parameters:
      - name: test
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Test
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - backup
      summary: Put Webdav
      operationId: put_webdav_api_v1_backup_webdav_put
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - backup
      summary: Post Webdav
      operationId: post_webdav_api_v1_backup_webdav_post
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/v1/backup/s3:
    get:
      tags:
      - backup
      summary: Get S3
      operationId: get_s3_api_v1_backup_s3_get
      security:
      - HTTPBearer: []
      parameters:
      - name: test
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Test
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - backup
      summary: Put S3
      operationId: put_s3_api_v1_backup_s3_put
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - backup
      summary: Post S3
      operationId: post_s3_api_v1_backup_s3_post
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...
components:
  schemas:
...
    Body_import_file_api_v1_backup_import_file_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        format:
          type: string
          title: Format
          default: json
        strategy:
          type: string
          title: Strategy
          default: skip_duplicate
        confirm_replace:
          type: boolean
          title: Confirm Replace
          default: false
      type: object
      required:
      - file
      title: Body_import_file_api_v1_backup_import_file_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    app__api__backup__ImportBody:
      properties:
        content:
          type: string
          title: Content
        format:
          type: string
          title: Format
          default: json
        strategy:
          type: string
          title: Strategy
          default: skip_duplicate
        confirm_replace:
          type: boolean
          title: Confirm Replace
          default: false
      type: object
      required:
      - content
      title: ImportBody
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer